backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==4.2.16
djangorestframework==3.15.2
idna==3.10
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
